Implemented Scipy sparse array to save track stats.
In theory, this should alleviate the need for limiting the maximum track duration as there is no need to preallocate large 2D dense arrays for storing the track stats variables. In practice, subsequent MCS identification and track stats files are still saved as dense array netCDF files for ease of use, as the number of MCS tracks are typically much smaller than total number of tracked clouds.

1. Modified trackstats_driver.py to use sparse array for storing the track statistics variables. Sparse array form of trackstats netCDF file is now the default file output. Dense array trackstats netCDF output is optional, with a flag added in the config file.
2. Modified identifymcs.py to use sparse array trackstats file. This change should significantly reduce the memory usage of the code.
3. Rewrote the code for matching merge/split clouds to MCS in identifymcs.py to make use of track start/end tracknumbers. A small bug was fixed where merge/split clouds are not saved for all time steps of an MCS. The new code should be more efficient and easier to understand.
4. Fixed a bug in trackstats_driver.py. After removing short-lived tracks and adjusting merge/split track numbers, the track start/end status must be updated to match the adjusted track numbers.