Fixed a rare bug in identifymcs.py.
The bug occurs in saving the merging/splitting CCS area to a given MCS track. In rare instances, the small merging/splitting cloud has 0 core/cold cloud area. I suspect that in sparse array form (csr_matrix), when doing "trackstat_corearea + trackstat_coldarea", the empty core/cold area data point is omitted. This causes a mismatch between the basetime and area array sizes, and error when they exceed the bounds.
Instead of adding the two csr_matrix arrays, the data for corearea and coldarea (as numpy arrays) are added, which seems to preserve the array size and fix the bug.