Added a fix to handle edge case.
Added a fix in pyflextrkr/trackstats_func.py to handle an edge case when more than 1 feature_number is found for a given tracknumber.
If this happens, the first feature_number is used, and a warning is displayed. It is unclear why this would occur, more test is needed. The issue is possibly in gettracks.py.